chore: rename LolexFakeTimers to ModernFakeTimers
