+++ packages/jest-fake-timers/package.json
@@ -11,14 +11,14 @@
   "types": "build/index.d.ts",
   "dependencies": {
     "@jest/types": "^26.0.0-alpha.0",
+    "@sinonjs/fake-timers": "^6.0.1",
     "jest-message-util": "^26.0.0-alpha.0",
     "jest-mock": "^26.0.0-alpha.0",
-    "jest-util": "^26.0.0-alpha.0",
-    "lolex": "^5.0.0"
+    "jest-util": "^26.0.0-alpha.0"
   },
   "devDependencies": {
-    "@types/lolex": "^5.1.0",
-    "@types/node": "*"
+    "@types/node": "*",
+    "@types/sinonjs__fake-timers": "^6.0.1"
   },
   "engines": {
     "node": ">= 10.14.2"
