+++ packages/jest-fake-timers/src/legacyFakeTimers.ts
